requirements.txt
scikit-learn
numpy
pdfplumber
python-docx
Flask
werkzeug
spacy
reportlab
